# social-media-automation/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Media Automation",
    "description": "Niche-agnostic AI content automation \u2014 generate + post videos to IG, FB, YouTube, TikTok, LinkedIn.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/topics": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Topics",
        "operationId": "list_topics_api_topics_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TopicRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "topics"
        ],
        "summary": "Create Topic",
        "description": "Add a topic manually (skips discovery).",
        "operationId": "create_topic_api_topics_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/run": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Run the pipeline end-to-end for a niche + topic, returns the created Post.",
        "operationId": "run_pipeline_action_api_posts_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/posts/{post_id}/regenerate": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Re-run the pipeline for an existing post (uses the same topic/niche).",
        "operationId": "regenerate_post_api_posts__post_id__regenerate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/post-now": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Post immediately to the named platforms, bypassing the schedule.",
        "operationId": "post_now_api_posts__post_id__post_now_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostNowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/topic-sources/{src_id}/run-now": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Manually trigger a topic source to discover topics now.",
        "description": "Synchronous discovery for a single source. Inserts new Topic rows.\n\nSkips topics whose content_hash already exists for this tenant (dedup).",
        "operationId": "run_topic_source_now_api_topic_sources__src_id__run_now_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "src_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Src Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules/{sched_id}/mark-done": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Manually mark a schedule as DONE (operator override).",
        "operationId": "mark_schedule_done_api_schedules__sched_id__mark_done_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sched_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sched Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic success/info envelope."
      },
...
      "Page_TopicRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopicRead"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[TopicRead]"
      },
...
      "PostNowRequest": {
        "properties": {
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Platforms",
            "description": "Subset of {instagram, facebook, youtube, tiktok, linkedin}"
          }
        },
        "type": "object",
        "required": [
          "platforms"
        ],
        "title": "PostNowRequest"
      },
      "PostNowResponse": {
        "properties": {
          "post_id": {
            "type": "integer",
            "title": "Post Id"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/PostNowResult"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "post_id",
          "attempts"
        ],
        "title": "PostNowResponse"
      },
...
      "PostRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "niche_id": {
            "type": "integer",
            "title": "Niche Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "video_length": {
            "type": "string",
            "title": "Video Length"
          },
          "video_format": {
            "type": "string",
            "title": "Video Format"
          },
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags"
          },
          "narrative_script": {
            "type": "string",
            "title": "Narrative Script"
          },
          "hook_text": {
            "type": "string",
            "title": "Hook Text"
          },
          "story_beats_json": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Story Beats Json"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "image_provider": {
            "type": "string",
            "title": "Image Provider"
          },
          "voice_provider": {
            "type": "string",
            "title": "Voice Provider"
          },
          "music_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music Provider"
          },
          "duration_sec": {
            "type": "number",
            "title": "Duration Sec"
          },
          "image_count": {
            "type": "integer",
            "title": "Image Count"
          },
          "media_cost_usd": {
            "type": "number",
            "title": "Media Cost Usd"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "error_log": {
            "type": "string",
            "title": "Error Log",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "id",
          "tenant_id",
          "niche_id",
          "status",
          "video_length",
          "video_format",
          "caption",
          "hashtags",
          "narrative_script",
          "hook_text",
          "story_beats_json",
          "llm_model",
          "image_provider",
          "voice_provider",
          "duration_sec",
          "image_count",
          "media_cost_usd"
        ],
        "title": "PostRead"
      },
...
      "RunPipelineRequest": {
        "properties": {
          "niche_id": {
            "type": "integer",
            "title": "Niche Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "manual_topic_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Topic Title"
          },
          "manual_topic_content": {
            "type": "string",
            "title": "Manual Topic Content",
            "default": ""
          },
          "video_length": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Length",
            "description": "short | long; overrides niche default"
          }
        },
        "type": "object",
        "required": [
          "niche_id"
        ],
        "title": "RunPipelineRequest"
      },
...
      "TopicCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "metadata_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Json"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TopicCreate",
        "description": "Manual topic \u2014 bypasses discovery sources."
      },
      "TopicRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tenant_id": {
            "type": "integer",
            "title": "Tenant Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "metadata_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Json"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "score_reason": {
            "type": "string",
            "title": "Score Reason",
            "default": ""
          },
          "suggested_angle": {
            "type": "string",
            "title": "Suggested Angle",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "used_for_post_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used For Post Id"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "id",
          "tenant_id",
          "content_hash",
          "title",
          "content",
          "metadata_json",
          "status"
        ],
        "title": "TopicRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}